oups, pas de suppression de rien
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@14514 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ _stable_/jeux/jeux_pipelines.php
@@ -159,9 +159,11 @@ function cron_jeux_cron($t){
 		$i++;
 	}
 	$sup.=(')');
-	spip_query('DELETE FROM spip_jeux WHERE `id_jeu` IN '.$sup);
-	spip_query('DELETE FROM spip_jeux_resultats WHERE `id_jeu` IN '.$sup);
-	spip_log('suppression jeux poubelle'.$sup);
+	if ($sup!='()') {
+		spip_query('DELETE FROM spip_jeux WHERE `id_jeu` IN '.$sup);
+		spip_query('DELETE FROM spip_jeux_resultats WHERE `id_jeu` IN '.$sup);
+		spip_log('suppression jeux poubelle'.$sup);
+		}
 	
 	}
